requirements.txt
astroid==4.0.4
certifi==2026.1.4
charset-normalizer==3.4.4
dill==0.4.1
dotenv==0.9.9
idna==3.11
iniconfig==2.3.0
isort==8.0.0
librt==0.8.1
mccabe==0.7.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
tomlkit==0.14.0
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
